Convert Movement (sliding) to an entity attribute.
I'm trying to reduce the instances of movement to a single instance per type, like a service, and instead apply movement against each entity based on it's type of movement. This warrants preserving the Movement (and eventually the collision strategy) as data inside of the local database. Which in turn requires converting the instances to accept entities to operate on.